moved ConfigEditor to components, parser for select objs
