+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/azure_monitor/response_parser.ts
@@ -108,7 +108,28 @@ export default class ResponseParser {
     return dimensions;
   }
 
-  static parseSubscriptions(result: any): Array<{ label: string; value: string }> {
+  static parseSubscriptions(result: any): Array<{ text: string; value: string }> {
+    const list: Array<{ text: string; value: string }> = [];
+
+    if (!result) {
+      return list;
+    }
+
+    const valueFieldName = 'subscriptionId';
+    const textFieldName = 'displayName';
+    for (let i = 0; i < result.data.value.length; i++) {
+      if (!_.find(list, ['value', _.get(result.data.value[i], valueFieldName)])) {
+        list.push({
+          text: `${_.get(result.data.value[i], textFieldName)} - ${_.get(result.data.value[i], valueFieldName)}`,
+          value: _.get(result.data.value[i], valueFieldName),
+        });
+      }
+    }
+
+    return list;
+  }
+
+  static parseSubscriptionsForSelect(result: any): Array<{ label: string; value: string }> {
     const list: Array<{ label: string; value: string }> = [];
 
     if (!result) {
